Separate pagination for better SRP
